Add Python wrapper pybuild11
The install operations have to be refined, but
the library works.

I have not written any unit test yet, but it is the next^2 step.
The next step is to add a method for the quaternion transform.

They are used by the cal.todmap.atm.py to create an atmospheric
observation performed by a whole focal plane.

Each process pick a pointing vector slightly shifted from the
pointings of the central detector. The correlation are ensured by the
same atmosphere that they observe, and that it was created previosly
with atm_sim->simulate();

That's it all, for now.