Some changes in the iterator type
1. Change were done on the iterator type. The `long` types were
   changed in `uint64_t`

2. The `observe` method of the calss `mpi_atm_sim` is now under
   maintenance. Strange issues were reported during the atmosphere
   observations. The TODs, in general, present some `bad data` and
   the function, during the observation operations, return a lot
   WARNINGS about an out-of-cone observations (not very often), and
   more often that not a "full index out of bounds" that it's
   not properly related to the observations but to the cells
   interpolation.

For the next future, the functions `mpi_atm_sim::observe` and
`mpi_atm_sim::interp` have to be double checked.

I want to remember that `cal` library was extracted by the TOAST
framework, and during this oparation a lot of features about the
processes distrubution were been lost. In this context, the
developers have to be used to work with these new framework and
they are filling every lack.