Modified `sim_tod.py` for implementing the smooth pointing
Now, the sping scan is modulated as a function of the AZ angle.
In this way the pointing are well distributed on the sky and are not
still concentrated on the two slab of sky at the edges of the observed sky

The improvements are significant, firt of all, the denoise algorithm
works a lot of order o magnitude better!