Spin tod (#25)
* Check on the az lims

In order to perform spin scans.
WARNING: A spin scan cannot be uniform. Too many hits at the edges
of the sky slab. The spin velocity has to be modulated accordingly
with the distance from the north pole.

* Update the simulation file

* Fixed some bugs in OpPointingHpix

Typo in init function

* Update the colorscale in plot_realization

update the single detector script

* Update Readme.md

* The opacity has been fixed (#22)

Now, the `atm_absorption_coefficient` uses the correct opacity
where the dry-air, O2 and H2O are taken into account, instead
the only H2O contribute.
 Changes to be committed:
	modified:   src/libcal/src/AATM_fun.cpp

* Add centos8 ohpc optimized singularity container definition file (#23)

Extended parallelization compatibility

* Add comments on total opacity

* Add the Focalplane class and tutorial notebooks

* Complete the atmospheric tutorial

* Delete Atmospheric_simulation_TOD.ipynb

* Update cookbook

* Update the cookbook

* Simulation pipeline for STRIP observations

* Scheduler for Slurm

* Completed the scheduler for indaco submission

* Feed the repository

* Modify the scheduler

* Update the Docker container

* Update the scheduler and the parameters files

* Add the two par files

* Update file par

* Update the scheduler

* Modified `sim_tod.py` for implementing the smooth pointing

Now, the sping scan is modulated as a function of the AZ angle.
In this way the pointing are well distributed on the sky and are not
still concentrated on the two slab of sky at the edges of the observed sky

The improvements are significant, firt of all, the denoise algorithm
works a lot of order o magnitude better!

* Elevation and azimuthal modulation

* Atm simulation modified

Co-authored-by: savaroskij <36457165+savaroskij@users.noreply.github.com>